move shared functions
